test: move tests into appropriate classes
