+++ vaadin-dashboard-flow-parent/vaadin-dashboard-flow-integration-tests/src/main/java/com/vaadin/flow/component/dashboard/tests/DashboardPage.java
@@ -16,7 +16,6 @@ import com.vaadin.flow.component.dashboard.DashboardSection;
 import com.vaadin.flow.component.dashboard.DashboardWidget;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.NativeButton;
-import com.vaadin.flow.component.html.Span;
 import com.vaadin.flow.router.Route;
 
 /**
@@ -30,7 +29,6 @@ public class DashboardPage extends Div {
 
         DashboardWidget widget1 = new DashboardWidget();
         widget1.setTitle("Widget 1");
-        widget1.setContent(new Div("Some content"));
         widget1.setId("widget-1");
 
         DashboardWidget widget2 = new DashboardWidget();
@@ -118,36 +116,6 @@ public class DashboardPage extends Div {
                 .ifPresent(dashboard::remove));
         removeFirstSection.setId("remove-first-section");
 
-        NativeButton addWidgetToFirstSection = new NativeButton(
-                "Add widget to first section");
-        addWidgetToFirstSection.addClickListener(
-                click -> getFirstSection(dashboard).ifPresent(section -> {
-                    DashboardWidget newWidget = new DashboardWidget();
-                    newWidget.setTitle("New widget");
-                    section.add(newWidget);
-                }));
-        addWidgetToFirstSection.setId("add-widget-to-first-section");
-
-        NativeButton removeFirstWidgetFromFirstSection = new NativeButton(
-                "Remove first widget from first section");
-        removeFirstWidgetFromFirstSection.addClickListener(
-                click -> getFirstSection(dashboard).ifPresent(section -> {
-                    List<DashboardWidget> currentWidgets = section.getWidgets();
-                    if (currentWidgets.isEmpty()) {
-                        return;
-                    }
-                    section.remove(currentWidgets.get(0));
-                }));
-        removeFirstWidgetFromFirstSection
-                .setId("remove-first-widget-from-first-section");
-
-        NativeButton removeAllFromFirstSection = new NativeButton(
-                "Remove all from first section");
-        removeAllFromFirstSection
-                .addClickListener(click -> getFirstSection(dashboard)
-                        .ifPresent(DashboardSection::removeAll));
-        removeAllFromFirstSection.setId("remove-all-from-first-section");
-
         NativeButton setMaximumColumnCount1 = new NativeButton(
                 "Set maximum column count 1");
         setMaximumColumnCount1
@@ -160,47 +128,9 @@ public class DashboardPage extends Div {
                 click -> dashboard.setMaximumColumnCount(null));
         setMaximumColumnCountNull.setId("set-maximum-column-count-null");
 
-        NativeButton increaseAllColspansBy1 = new NativeButton(
-                "Increase all colspans by 1");
-        increaseAllColspansBy1.addClickListener(click -> dashboard.getWidgets()
-                .forEach(widget -> widget.setColspan(widget.getColspan() + 1)));
-        increaseAllColspansBy1.setId("increase-all-colspans-by-1");
-
-        NativeButton decreaseAllColspansBy1 = new NativeButton(
-                "Decrease all colspans by 1");
-        decreaseAllColspansBy1.addClickListener(click -> dashboard.getWidgets()
-                .forEach(widget -> widget.setColspan(widget.getColspan() - 1)));
-        decreaseAllColspansBy1.setId("decrease-all-colspans-by-1");
-
-        NativeButton updateContentOfTheFirstWidget = new NativeButton(
-                "Update content of the first widget");
-        updateContentOfTheFirstWidget.addClickListener(click -> {
-            List<DashboardWidget> widgets = dashboard.getWidgets();
-            if (!widgets.isEmpty()) {
-                widgets.get(0).setContent(new Span("Updated content"));
-            }
-        });
-        updateContentOfTheFirstWidget
-                .setId("update-content-of-the-first-widget");
-
-        NativeButton removeContentOfTheFirstWidget = new NativeButton(
-                "Remove content of the first widget");
-        removeContentOfTheFirstWidget.addClickListener(click -> {
-            List<DashboardWidget> widgets = dashboard.getWidgets();
-            if (!widgets.isEmpty()) {
-                widgets.get(0).setContent(null);
-            }
-        });
-        removeContentOfTheFirstWidget
-                .setId("remove-content-of-the-first-widget");
-
         add(addMultipleWidgets, removeFirstAndLastWidgets, removeAll,
                 addSectionWithMultipleWidgets, removeFirstSection,
-                addWidgetToFirstSection, removeFirstWidgetFromFirstSection,
-                removeAllFromFirstSection, updateContentOfTheFirstWidget,
-                removeContentOfTheFirstWidget, setMaximumColumnCount1,
-                setMaximumColumnCountNull, increaseAllColspansBy1,
-                decreaseAllColspansBy1, dashboard);
+                setMaximumColumnCount1, setMaximumColumnCountNull, dashboard);
     }
 
     private static Optional<DashboardSection> getFirstSection(
